pyproject.toml
[project]
name = "auge-transportation-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.12.0",
    "asyncpg>=0.30.0",
    "crewai>=0.126.0",
    "crewai-tools>=0.47.1",
    "fastapi>=0.115.9",
    "langchain-ollama>=0.3.3",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
]
